Harden Smart Hub quota surfaces
Add Smart Display repair state and bridge recovery paths across macOS, core, and mobile.

Improve quota normalization, provider branding, Android pulse surfaces, launcher assets, and Firestore parsing coverage.

Update Smart Display and Pixel Clock docs with the new behavior.

+++ android/app/src/main/java/com/openburnbar/data/stores/ProviderDashboardStore.kt
@@ -14,7 +14,7 @@ data class ProviderDashboard(
     val providerSummaries: List<RollupSummary> = emptyList(),
     val quotaSnapshots: List<ProviderQuotaSnapshot> = emptyList(),
     val totalCost: Double = 0.0,
-    val totalTokens: Int = 0
+    val totalTokens: Long = 0
 )
 
 class ProviderDashboardStore(
@@ -40,7 +40,7 @@ class ProviderDashboardStore(
                     providerSummaries = rollups.providerSummaries,
                     quotaSnapshots = quotas,
                     totalCost = rollups.allTime,
-                    totalTokens = rollups.totals["tokens"]?.toInt() ?: 0
+                    totalTokens = rollups.totals["tokens"]?.toLong() ?: 0L
                 )
             } catch (e: Exception) {
                 _error.value = e.message
